not-required/agent-workflow/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart

# Server-Sent Events for Streaming
sse-starlette

# Data Models
pydantic

# Vector Database
chromadb

# LLM Integration
groq

# Environment Variables
python-dotenv

requests
numpy
typing-extensions
vosk
soundfile
twilio
websockets
aiohttp

# Hybrid retrieval and summarization
whoosh
transformers  # optional but recommended for better summarization
sentence-transformers  # ensure embedder compatibility
pyngrok
 
# Optional OCR / scraping helpers (recommended for full ingestion features)
pdf2image
opencv-python-headless
beautifulsoup4
aiohttp
